Added tests and updated dispatcher to make it easier to create new callback execution patterns
